Move topic video embed into embeds module
