+++ .gitignore
@@ -3,7 +3,5 @@
 !.env.example
 *.pyc
 media/
-!semanticnews/topics/utils/media/
-!semanticnews/topics/utils/media/**
 __pycache__/
 !.env.example
