+++ semanticnews/settings.py
@@ -49,7 +49,6 @@ INSTALLED_APPS = [
     'semanticnews.topics.utils.narratives',
     'semanticnews.topics.utils.mcps',
     'semanticnews.topics.utils.images',
-    'semanticnews.topics.utils.media',
     'semanticnews.topics.utils.embeds',
     'semanticnews.topics.utils.relations',
     'semanticnews.topics.utils.data',
