+++ semanticnews/topics/templates/topics/topics_detail.html
@@ -124,7 +124,7 @@
 
         {% include "topics/narratives/card.html" %}
 
-        {% include "topics/media/card.html" %}
+        {% include "topics/embeds/card.html" %}
 
         {% include "topics/relations/card.html" %}
 
@@ -215,7 +215,7 @@
     {% include "topics/data/scripts.html" %}
     {% include "topics/narratives/scripts.html" %}
     {% include "topics/images/scripts.html" %}
-    {% include "topics/media/scripts.html" %}
+    {% include "topics/embeds/scripts.html" %}
     {% include "topics/relations/scripts.html" %}
     {% include "topics/timeline/scripts.html" %}
     {% include "topics/mcps/scripts.html" %}
