+++ semanticnews/topics/utils/embeds/templates/topics/embeds/scripts.html
@@ -1,0 +1,2 @@
+{% load static %}
+<script src="{% static 'topics/embeds/topic_embeds.js' %}"></script>
