--- semanticnews/topics/utils/media/admin.py
@@ -1,9 +1,0 @@
-from django.contrib import admin
-
-from .models import TopicYoutubeVideo
-
-
-@admin.register(TopicYoutubeVideo)
-class TopicYoutubeVideoAdmin(admin.ModelAdmin):
-    list_display = ("topic", "url", "title", "created_at")
-    search_fields = ("topic__title", "url", "title")
