--- semanticnews/topics/utils/media/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class MediaConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'semanticnews.topics.utils.media'
-    app_label = 'topics'
