--- semanticnews/topics/utils/media/templates/topics/media/scripts.html
@@ -1,2 +1,0 @@
-{% load static %}
-<script src="{% static 'topics/media/topic_media.js' %}"></script>
